requirements.CUDA.txt
keras==2.4.3
onnx==1.8.1
onnxruntime-gpu==1.7.0
netron==4.8.3
tf2onnx==1.8.4
